+++ test/unit/evvm/correct/unitTestCorrect_Core_disperseCaPay.t.sol
@@ -21,9 +21,7 @@ import "test/Constants.sol";
 import "@evvm/testnet-contracts/library/Erc191TestBuilder.sol";
 
 import {Core} from "@evvm/testnet-contracts/contracts/core/Core.sol";
-import {
-    CoreError
-} from "@evvm/testnet-contracts/library/errors/CoreError.sol";
+import {CoreError} from "@evvm/testnet-contracts/library/errors/CoreError.sol";
 
 contract unitTestCorrect_Core_disperseCaPay is Test, Constants {
     //function executeBeforeSetUp() internal override {}
@@ -106,9 +104,7 @@ contract unitTestCorrect_Core_disperseCaPay is Test, Constants {
         );
     }
 
-    function test__unit_correct__disperseCaPay__denyList()
-        external
-    {
+    function test__unit_correct__disperseCaPay__denyList() external {
         vm.startPrank(ADMIN.Address);
         core.proposeListStatus(0x02); // Activate denyList
         skip(1 days + 1); // Skip timelock
@@ -146,9 +142,7 @@ contract unitTestCorrect_Core_disperseCaPay is Test, Constants {
         );
     }
 
-    function test__unit_correct__disperseCaPay__allowList()
-        external
-    {
+    function test__unit_correct__disperseCaPay__allowList() external {
         vm.startPrank(ADMIN.Address);
         core.proposeListStatus(0x01); // Activate allowList
         skip(1 days + 1); // Skip timelock
@@ -184,6 +178,57 @@ contract unitTestCorrect_Core_disperseCaPay is Test, Constants {
             core.getBalance(address(this), PRINCIPAL_TOKEN_ADDRESS),
             0,
             "ca dosent recieve rewards because is no an staker"
+        );
+    }
+
+    function test__unit_correct__disperseCaPay__RewardFlowDistribution_false()
+        external
+    {
+        uint256 currentSupply = core.getCurrentSupply();
+        uint256 totalSupply = core.getEvvmMetadata().totalSupply;
+        uint256 remainingSupply = totalSupply - currentSupply;
+        core.addBalance(
+            address(this),
+            PRINCIPAL_TOKEN_ADDRESS,
+            remainingSupply - 1
+        );
+
+        vm.startPrank(ADMIN.Address);
+        core.proposeChangeRewardFlowDistribution();
+        skip(1 days);
+        core.acceptChangeRewardFlowDistribution();
+        vm.stopPrank();
+
+        (uint256 amount) = _addBalance(address(this), ETHER_ADDRESS, 0.1 ether);
+
+        core.setPointStaker(address(this), 0x01);
+
+        CoreStructs.DisperseCaPayMetadata[]
+            memory toData = new CoreStructs.DisperseCaPayMetadata[](1);
+
+        toData[0] = CoreStructs.DisperseCaPayMetadata({
+            amount: amount,
+            toAddress: COMMON_USER_NO_STAKER_2.Address
+        });
+
+        core.disperseCaPay(toData, ETHER_ADDRESS, amount);
+
+        assertEq(
+            core.getBalance(COMMON_USER_NO_STAKER_2.Address, ETHER_ADDRESS),
+            amount,
+            "Amount should be recibed"
+        );
+
+        assertEq(
+            core.getBalance(address(this), ETHER_ADDRESS),
+            0,
+            "Amount should be deducted"
+        );
+
+        assertEq(
+            core.getBalance(address(this), PRINCIPAL_TOKEN_ADDRESS),
+            remainingSupply - 1,
+            "Staker ca should not recieve rewards because reward flow distribution is false"
         );
     }
 }